print arr changed
