+++ makefile
@@ -1,50 +1,55 @@
-MAKEFLAGS += -s
+NAME = cub3d
 
-CC				= cc
-FLAGS			= -Wall -Wextra -Werror -g -w
-HDRS			= includes
-MLX_HDRS		= mlx
-LIBFT_DIR		= libft
-LIBFT			= $(LIBFT_DIR)/libft.a
-SRCS_DIR		= ./
-OBJ_DIR			= obj
+INCLUDE = includes
+SRC_DIR = src
+OBJ_DIR = objs
 
-SRCS			= $(shell find $(SRCS_DIR) -type f -name '*.c' ! -path '$(SRCS_DIR)/mlx/*' ! -path '$(SRCS_DIR)/libft/*')
-SRC_O			= $(SRCS:$(SRCS_DIR)%.c=$(OBJ_DIR)/%.o)
+LIBFT_DIR = libft
+LIBFT = $(LIBFT_DIR)/libft.a
 
-MLX_FLAGS		= -L mlx -lmlx -lXpm -lXext -lX11 -lm
-LIB_PATH		= mlx
-LIB_MLX			= $(LIB_PATH)/libmlx.a
+MLX_DIR = mlx
+MLX = $(MLX_DIR)/libmlx_Linux.a
 
-NAME			= cub3d
-RM				= rm -f
-CYAN			= "\033[36m"
-RED				= "\033[31m"
-GREEN			= "\033[32m"
-RESET			= "\033[0m"
+CC = cc
+FLAGS = -Wall -Werror -g
+MLX_FLAGS = -lX11 -lXext -lm
+
+FILES = main.c\
+		free.c\
+		printing.c\
+		treat_map.c\
+		treat_file.c\
+		player_check.c
+
+OBJS = $(FILES:%.c=$(OBJ_DIR)/%.o)
+
+RM = rm -f
+
+CYAN = "\033[36m"
+RED    = "\033[31m"
+GREEN = "\033[32m"
+RESET = "\033[0m"
 
 all: $(NAME)
 
-$(NAME): $(LIBFT) $(SRC_O) $(LIB_MLX)
+$(NAME): $(LIBFT) $(MLX) $(OBJS)
 	@echo $(CYAN)"Compilation de Cub3D..."$(RESET)
-	@$(CC) $(FLAGS) $(SRC_O) $(LIB_MLX) $(MLX_FLAGS) -L libs -lft -I includes -I libft -I libft/includes -I $(MLX_HDRS) -o $(NAME)
+	@$(CC) $(FLAGS) $(OBJS) $(LIBFT) $(MLX) -I $(INCLUDE) -I $(MLX_DIR) -I $(LIBFT_DIR) -I $(LIBFT_DIR)/includes -o $(NAME) $(MLX_FLAGS)
 	@echo $(GREEN)"Exécutable $(NAME) créé !"$(RESET)
 
 $(LIBFT):
-	@mkdir -p libs
+	@echo -n "Compiling libft..."
+	@make -C $(LIBFT_DIR) > /dev/null
+	@echo "\nLibft compilation done!"
+
+$(MLX):
+	@echo -n "Compiling libmlx..."
+	@make -C $(MLX_DIR) > /dev/null
+	@echo "\nLibmlx compilation done!"
+
+$(OBJ_DIR)/%.o: $(SRC_DIR)/%.c
 	@mkdir -p $(OBJ_DIR)
-	$(MAKE) -C $(LIBFT_DIR)
-	@mv $(LIBFT) libs
-
-$(OBJ_DIR)/%.o: $(SRCS_DIR)%.c includes/*.h libft/*.h libft/includes/*.h
-	@mkdir -p $(dir $@)
-	@$(CC) $(FLAGS) -I includes -I libft -I libft/includes -I $(MLX_HDRS) -c $< -o $@
-
-# Règle pour la compilation de MiniLibX
-$(LIB_MLX):
-	@echo $(CYAN)"Compilation de MiniLibX..."$(RESET)
-	@make -s -C $(LIB_PATH)
-	@echo $(GREEN)"MiniLibX compilée !"$(RESET)
+	@$(CC) $(FLAGS) -I $(INCLUDE) -I $(LIBFT_DIR) -I $(LIBFT_DIR)/includes -I $(MLX_DIR) -c $< -o $@
 
 clean:
 	@echo $(CYAN)"Suppression des fichiers objets..."$(RESET)
@@ -52,10 +57,10 @@ clean:
 	$(MAKE) -C $(LIBFT_DIR) clean
 
 fclean: clean
-	@echo $(CYAN)"Suppression de l'exécutable..."$(RESET)
+	@echo "Suppression de l'exécutable..."
 	@$(RM) $(NAME)
-	@echo $(CYAN)"Suppression des libs..."$(RESET)
-	@$(RM) libs/*.a
+	@echo "Suppression de libft..."
+	@make -C $(LIBFT_DIR) fclean > /dev/null
 
 re: fclean all
