+++ src/player_check.c
