+++ test.cub
@@ -8,7 +8,3 @@ C 51, 4, 43
    1111 111
 11111W111
  111111
-
-1111111111
-11111W111
- 111111
